加入对图像大小的判断，显示时如果图像过小，则不缩放直接显示。

+++ OpenCV-Cpp-2.4.9/CVAlgorithm/src/app/widgets/imagegridwidget.h
@@ -33,8 +33,8 @@ private:
     int m_maxColumns;
 
     QTimer* m_resizeTimer = new QTimer(this);
-    int idealItemWidth{-1};
-    int idealItemHeight{-1};
+    double idealItemWidth{-1};
+    double idealItemHeight{-1};
 };
 
 #endif // IMAGEGRIDWIDGET_H
